requirements.txt
Django==4.2.23
djangorestframework==3.16.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
